requirements.txt
requests>=2.31.0
python-slugify>=8.0.1
Pillow>=10.0.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
